+++ src/Schema/Column/ColumnBuilder.php
@@ -271,4 +271,15 @@ class ColumnBuilder
     {
         return new JsonColumn(ColumnType::JSON);
     }
+
+    /**
+     * Builds a column with the abstract type `enum`.
+     *
+     * @param string[]|null $values The list of possible values for the enum column.
+     * @param string|null $dbType The database type of the column. When null, the column will use a CHECK constraint.
+     */
+    public static function enum(?array $values = null, ?string $dbType = null): EnumColumn
+    {
+        return new EnumColumn(dbType: $dbType, values: $values);
+    }
 }