+++ tests/Provider/ColumnDefinitionParserProvider.php
@@ -30,27 +30,35 @@ class ColumnDefinitionParserProvider
             ['int CHECK (value > (1 + 5))', ['type' => 'int', 'check' => 'value > (1 + 5)']],
             [
                 "enum('a','b','c')",
-                ['type' => 'enum', 'enumValues' => ['a', 'b', 'c']],
+                ['type' => 'enum', 'values' => ['a', 'b', 'c']],
             ],
             [
                 "enum('a','b','c') NOT NULL",
-                ['type' => 'enum', 'enumValues' => ['a', 'b', 'c'], 'notNull' => true],
+                ['type' => 'enum', 'values' => ['a', 'b', 'c'], 'notNull' => true],
+            ],
+            'enum-upper-case' => [
+                "ENUM('a','b','c') NOT NULL",
+                ['type' => 'enum', 'values' => ['a', 'b', 'c'], 'notNull' => true],
+            ],
+            'enum-custom-case' => [
+                "eNUm('a','b','c') NOT NULL",
+                ['type' => 'enum', 'values' => ['a', 'b', 'c'], 'notNull' => true],
             ],
             'enum-with-square-brackets' => [
                 "enum('[one]', 'the [two]', 'the [three] to') NOT NULL",
-                ['type' => 'enum', 'enumValues' => ['[one]', 'the [two]', 'the [three] to'], 'notNull' => true],
+                ['type' => 'enum', 'values' => ['[one]', 'the [two]', 'the [three] to'], 'notNull' => true],
             ],
             'enum-with-parentheses' => [
                 "enum('(one)', 'the (two)', 'the (three) to') NOT NULL",
-                ['type' => 'enum', 'enumValues' => ['(one)', 'the (two)', 'the (three) to'], 'notNull' => true],
+                ['type' => 'enum', 'values' => ['(one)', 'the (two)', 'the (three) to'], 'notNull' => true],
             ],
             'enum-with-escaped-quotes' => [
                 "enum('hello''world''', 'the ''[feature]''') NOT NULL",
-                ['type' => 'enum', 'enumValues' => ["hello'world'", "the '[feature]'"], 'notNull' => true],
+                ['type' => 'enum', 'values' => ["hello'world'", "the '[feature]'"], 'notNull' => true],
             ],
             'enum-with-parentheses-and-escaped-quotes' => [
                 "enum('''hey (one)', 'the (t''wo)', 'the (three) ''to') NOT NULL",
-                ['type' => 'enum', 'enumValues' => ['\'hey (one)', 'the (t\'wo)', 'the (three) \'to'], 'notNull' => true],
+                ['type' => 'enum', 'values' => ['\'hey (one)', 'the (t\'wo)', 'the (three) \'to'], 'notNull' => true],
             ],
             ['int[]', ['type' => 'int', 'dimension' => 1]],
             ['string(126)[][]', ['type' => 'string', 'size' => 126, 'dimension' => 2]],