+++ CHANGELOG.md
@@ -178,6 +178,7 @@
 - Chg #1103: Remove `AbstractCommand::refreshTableSchema()` method (@vjik)
 - Chg #1106: Remove parameters from `PdoConnectionInterface::getActivePdo()` method (@vjik)
 - Bug #1109: Fix column definition parsing in cases with brackets and escaped quotes (@vjik)
+- New #1107: Add abstract enumeration column type (@vjik)
 
 ## 1.3.0 March 21, 2024